feat(Weapon): re-work
