+++ src/main/java/github/xnzvl/karak/items/Item.java
@@ -2,7 +2,6 @@ package github.xnzvl.karak.items;
 
 import github.xnzvl.karak.items.chests.Chest;
 import github.xnzvl.karak.items.spells.Spell;
-import github.xnzvl.karak.items.weapons.Weapon;
 import github.xnzvl.karak.utils.Either;
 
 import java.util.function.Consumer;
