--- src/main/java/github/xnzvl/karak/items/weapons/impl/Axe.java
@@ -1,18 +1,0 @@
-package github.xnzvl.karak.items.weapons.impl;
-
-import github.xnzvl.karak.items.weapons.VariousWeapon;
-
-/**
- * Wielder should gain +3 power.
- *
- * @author Jakub Nezval
- */
-public class Axe extends VariousWeapon {
-    public Axe() {
-        super(
-                "Axe",
-                "+3 power.",
-                3
-        );
-    }
-}
