--- src/main/java/github/xnzvl/karak/items/weapons/impl/Dagger.java
@@ -1,18 +1,0 @@
-package github.xnzvl.karak.items.weapons.impl;
-
-import github.xnzvl.karak.items.weapons.VariousWeapon;
-
-/**
- * Wielder should gain +1 power.
- *
- * @author Jakub Nezval
- */
-public class Dagger extends VariousWeapon {
-    public Dagger() {
-        super(
-                "Dagger",
-                "+1 power.",
-                1
-        );
-    }
-}
