+++ src/main/java/github/xnzvl/karak/strengthfuls/Monster.java
@@ -3,13 +3,11 @@ package github.xnzvl.karak.strengthfuls;
 import github.xnzvl.karak.description.Description;
 import github.xnzvl.karak.items.Item;
 import github.xnzvl.karak.items.Key;
+import github.xnzvl.karak.items.Weapon;
 import github.xnzvl.karak.items.chests.impl.OrdinaryChest;
 import github.xnzvl.karak.items.chests.impl.RubyChest;
 import github.xnzvl.karak.items.spells.impl.FlamingSword;
 import github.xnzvl.karak.items.spells.impl.HealingWarp;
-import github.xnzvl.karak.items.weapons.impl.Axe;
-import github.xnzvl.karak.items.weapons.impl.Dagger;
-import github.xnzvl.karak.items.weapons.impl.Sword;
 
 /**
  * Enum for monster units.
@@ -17,12 +15,12 @@ import github.xnzvl.karak.items.weapons.impl.Sword;
  * @author Jakub Nezval
  */
 public enum Monster implements Description, Strength {
-    RAT                 ("Rat", "", 5, Item.from(new Dagger())),
+    RAT                 ("Rat", "", 5, Item.from(Weapon.DAGGER)),
     SPIDER              ("Spider", "", 6, Item.from(new HealingWarp())),
     SKELETON_KEY_KEEPER ("Skeleton Key Keeper", "", 7, Item.from(new Key())),
     MUMMY               ("Mummy", "", 7, Item.from(new FlamingSword())),
-    SKELETON_SWORDSMAN  ("Skeleton Swordsman", "", 9, Item.from(new Sword())),
-    SKELETON_KING       ("Skeleton King", "", 10, Item.from(new Axe())),
+    SKELETON_SWORDSMAN  ("Skeleton Swordsman", "", 9, Item.from(Weapon.SWORD)),
+    SKELETON_KING       ("Skeleton King", "", 10, Item.from(Weapon.AXE)),
     FALLEN              ("Fallen", "", 12, Item.from(new OrdinaryChest(false))),
     DRAGON              ("Dragon", "", 15, Item.from(new RubyChest()));
